README modification, test files ordered
