+++ src/_tests_/HeaderPrincipal/HeaderPrincipal.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import HeaderPrincipal from './HeaderPrincipal';
+import HeaderPrincipal from '../../components/HeaderPrincipal/HeaderPrincipal';
 
 it('should render header', () => {
     const { getByRole, debug } = render(<HeaderPrincipal />)
